refactor(move-package): 패키지 이동
테스트 패키지에서 소스코드 패키지로 도메인을 구성하는 객체들을 이동시켰습니다.
