+++ src/main/java/domain/bet/BetProfit.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.bet;
 
 import domain.enums.GameResult;
 import domain.participant.Name;
@@ -8,8 +8,8 @@ import java.util.Map;
 import java.util.Set;
 
 public class BetProfit {
-    private final static int ZERO = 0;
-    private final static double BLACKJACK_YIELD = 1.5;
+    private static final int ZERO = 0;
+    private static final double BLACKJACK_YIELD = 1.5;
 
     private final Map<Name, Integer> betProfit = new HashMap<>();
