+++ src/test/java/domain/BetTest.java
@@ -3,6 +3,7 @@ package domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
 
+import domain.bet.Bet;
 import domain.enums.GameResult;
 import domain.participant.Name;
 import java.util.List;
